requirements.txt
Flask
requests
python-dotenv
google-generativeai
streamlit 
pillow
huggingface_hub
jsonify